login logout

+++ backend/controllers/chatController.js
@@ -1,9 +1,10 @@
 import { pool } from "../db.js";
 
-export const chatExists = async (chatId) => {
+export const chatExists = async (chatId, userId) => {
+  
   const chatCheck = await pool.query(
-      "SELECT title FROM chats WHERE id = $1",
-      [chatId]
+      "SELECT title FROM chats WHERE id = $1 and user_id=$2",
+      [chatId, userId]
     );
 
   if (chatCheck.rows.length > 0) {
@@ -21,7 +22,8 @@ export const chatExists = async (chatId) => {
 
 export const getChatsTitles = async (req, res) => {
   try {
-    const result = await pool.query("SELECT id, title FROM chats");
+    const userId = req.userId;
+    const result = await pool.query("SELECT id, title FROM chats where user_id=$1", [userId]);
     return res.json(result.rows);  // завжди повертає валідний JSON
   } catch (error) {
     console.error("Database error fetching chat titles:", error.message);
@@ -33,8 +35,8 @@ export const changeChatTitle = async(req, res) => {
   try {
     const {title} = req.body;
     const {id} = req.params;
-
-    const result = await pool.query(`UPDATE chats SET title=$1 where id=$2 RETURNING *`,  [title, id]);
+    const userId = req.userId;
+    const result = await pool.query(`UPDATE chats SET title=$1 where id=$2 and user_id=$3 RETURNING *`,  [title, id, userId]);
 
     if (result.rowCount === 0){
       return res.status(404).json({ error: "Chat not found" });
@@ -50,10 +52,10 @@ export const changeChatTitle = async(req, res) => {
 export const createChat = async(req, res) => {
   try {
     const {id, title, created_at} = req.body;
-
-    const result = await pool.query(`INSERT INTO chats(id, title, created_at) VALUES(
-        $1, $2, $3::timestamptz
-        ) RETURNING *`,  [id, title, created_at]);
+    const userId = req.userId;
+    const result = await pool.query(`INSERT INTO chats(id, title, created_at, user_id) VALUES(
+        $1, $2, $3::timestamptz, $4
+        ) RETURNING *`,  [id, title, created_at, userId]);
 
     return res.json(result.rows[0]);
 
@@ -65,9 +67,10 @@ export const createChat = async(req, res) => {
 
 export const getLastChats = async(req, res) => {
   try {
+    const userId = req.userId;
     const result = await pool.query(`SELECT * FROM chats
-    WHERE created_at >= CURRENT_DATE - INTERVAL '7 days'
-    ORDER BY created_at DESC;`);
+    WHERE created_at >= CURRENT_DATE - INTERVAL '7 days' and user_id=$1
+    ORDER BY created_at DESC;`, [userId]);
 
     return res.json(result.rows);
 
@@ -80,7 +83,9 @@ export const getLastChats = async(req, res) => {
 export const deleteChat = async(req, res) => {
   try{
     const {id} = req.params;
-    const result = await pool.query("DELETE FROM chats WHERE id=$1 RETURNING *", [id]);
+    const userId = req.userId;
+    const result = await pool.query("DELETE FROM chats WHERE id=$1 and user_id=$2 RETURNING *", [id, userId]);
+    
 
     const resultMsg = await pool.query("DELETE FROM messagesTable WHERE chat_id=$1 RETURNING *", [id]);
 
@@ -98,12 +103,12 @@ export const deleteChat = async(req, res) => {
 export const findChatBySearch = async(req, res) => {
   try {
     const { search } = req.body;
-
+  const userId = req.userId;  
     const result = await pool.query(`select chats.id, chats.title, chats.created_at from chats 
       join messagesTable on chats.id=messagesTable.chat_id 
-      where chats.title ilike $1 or messagesTable.message ilike $1 
+      where chats.user_id=$1 and (chats.title ilike $2 or messagesTable.message ilike $2)
       GROUP BY chats.id, chats.title, chats.created_at
-        ORDER BY chats.created_at DESC`, [`%${search}%`]);
+        ORDER BY chats.created_at DESC`, [userId,`%${search}%`]);
 
     return res.json(result.rows);
   } catch (error) {
@@ -115,8 +120,9 @@ export const findChatBySearch = async(req, res) => {
 
 export const checkingChat = async(req, res) => {
     const { id: chatId } = req.params;
+    const userId = req.userId;  
   try {
-    const result = await chatExists(chatId);
+    const result = await chatExists(chatId, userId);
 
     if (!result || typeof result !== "object") {
       return res.status(500).json({ error: "Invalid result from chatExists" });
